Add an "all" generator as a meta generator using others.
This generator will utilize the other existing generators to generate a
random name from any one of them.

I'm not 100% convinced of the API yet so I may make some changes to it.
It doesn't feel very future-proof in how it constructs other generators
itself.  Maybe that should be done outside of the class or in a static
method?