Add __toString() to all generators.
Reduce redundancy of code using random-name-generator by providing a
`__toString()` method to alias the `getName` method.  Rather than repeat
this method on all classes, this uses an abstract class to add the
behavior since all generators would theoreticaly be cast to string in
the same way.

This does not add the __toString method to the interface yet as that
would break compatibility to downstream libraries.  Version 3.0 will
have __toString added as part of the interface.